Remove const on Context in ResultRetriever
This change is to allow modifying the context on deserialization of
certain special file types related to C++20 modules
(such as DDI files or similar).

We sometimes need to use the data from these files to reconstruct
information about the compilation context which cannot be obtained
solely through the command-line arguments and which would otherwise
require running the compiler and observing the output.

Specifically this is a problem when trying to determine the correct
output name for the BMI file.

GCC does not allow to specify the BMI file-name, and the compiler
chosen BMI name depends on the name (identifier) of the C++20 module in
the translation unit.

Clang and MSVC allow the output BMI file-name to be optionally
specified but if it is not specified there is no guarantee or
fixed schema for how the name will be chosen.